+++ app/src/main/java/csis3175/w23/g11/rooftown/PostingFragment.java
@@ -1,4 +1,4 @@
-package csis3175.proj;
+package csis3175.w23.g11.rooftown;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -9,16 +9,12 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
-public class ProfileFragment extends Fragment {
-
-    public static ProfileFragment newInstance() {
-        return new ProfileFragment();
-    }
+public class PostingFragment extends Fragment {
 
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
                              @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_profile, container, false);
+        return inflater.inflate(R.layout.fragment_posting, container, false);
     }
 
 }
